Modified to support RFD868x
When using RFD868x radios for both telemetry and RC control (via PPM passthrough), RSSI is only reported via the MAVLink RADIO_STATUS message (identical structure to RADIO message).

This commit modifies the code so that RSSI is no longer read from the RC_CHANNELS message, but from the RADIO_STATUS message instead.

Additionally, RADIO_STATUS message is only output if a HEARTBEAT messgae is received (in to the radio from GCS). If no GCS is connected via the Aux port then no RADIO_STATUS mesage will be received (therefore no RSSI). To overcome this, added in the option for the Teensy to output a "dummy" HEARTBEAT message if no GCS is connected.